requirements.txt
gradio_client==1.5.2
Pillow==11.0.0
playsound==1.3.0
protobuf==5.29.2
SpeechRecognition==3.12.0
transformers==4.47.1
torch==2.5.1
google-generativeai

# pip install -U google-generativeai if needed